+++ src/js/item/factory.js
@@ -1,27 +1,33 @@
+import AudioItem from "./audioItem";
 import PictureItem from "./pictureItem";
+import ProviderItem from "./providerItem";
 import TextItem from "./textItem";
+import VideoItem from "./videoItem";
 
 const classes = {
+    AudioItem,
     PictureItem,
-    TextItem
+    ProviderItem,
+    TextItem,
+    VideoItem
 };
 
 export default class Factory {
     /**
      * @param {string} name
-     * @param {object} args
-     * @returns {TextItem|PictureItem}
+     * @param {object} parameters
+     * @returns {AudioItem|PictureItem|ProviderItem|TextItem|VideoItem}
      */
-    create(name, args) {
+    create(name, parameters) {
         let className = name.toLowerCase() + 'Item';
         className = className.charAt(0).toUpperCase() + className.slice(1);
 
         try {
-            return new classes[className](args);
+            return new classes[className](parameters);
         } catch (exception) {
             let message;
-            if ("undefined" !== typeof exception.name && exception.name === 'TypeError') {
-                message = 'Invalid item type "' + name + '" (allowed values: "text", "picture")';
+            if ('undefined' !== typeof exception.name && exception.name === 'TypeError') {
+                message = 'Invalid item type "' + name + '" (allowed values: "audio", "picture", "provider", "text", "video")';
             } else {
                 message = exception.message;
             }