+++ src/js/item/pictureItem.js
@@ -21,7 +21,7 @@ export default class PictureItem extends BaseItem {
      * @returns {HTMLElement}
      */
     createPicture() {
-        const element = this.domHelper.createElement('img', 'picture');
+        const element = this.domHelper.createElement('img', {'class': 'picture'});
         element.src = this.path;
 
         if ('undefined' !== typeof this.caption) {
@@ -38,7 +38,7 @@ export default class PictureItem extends BaseItem {
      */
     renderHtml() {
         const element = this.createItemElement();
-        const pictureItem =  this.domHelper.createElement('div', 'picture-item');
+        const pictureItem = this.domHelper.createElement('div', {'class': 'picture-item'});
 
         if ('undefined' !== typeof this.caption) {
             pictureItem.setAttribute('data-caption', this.caption);